backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
opencv-python==4.8.1.78
mediapipe==0.10.31
numpy==1.26.4
Pillow==10.1.0
python-multipart==0.0.6
